VBL_API/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coremltools==9.0
cycler==0.12.1
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
pyaml==25.7.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
textblob==0.19.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.40.0
virtualenv==20.35.4
